refactor: move feature strategies store
